change structure to easy adopt new dbms
