+++ src/snowflake/snowpark/_internal/data_source/dbms_dialects/base_dialect.py
@@ -1,0 +1,15 @@
+#
+# Copyright (c) 2012-2025 Snowflake Computing Inc. All rights reserved.
+#
+
+from abc import ABC, abstractmethod
+from snowflake.snowpark.types import StructType
+import logging
+
+logger = logging.getLogger(__name__)
+
+
+class BaseDialect(ABC):
+    @abstractmethod
+    def generate_select_query(self, table_or_query: str, schema: StructType) -> str:
+        pass
