+++ src/snowflake/snowpark/dataframe_writer.py
@@ -25,9 +25,9 @@ from snowflake.snowpark._internal.ast.utils import (
     DATAFRAME_AST_PARAMETER,
     build_table_name,
 )
-from snowflake.snowpark._internal.data_source_utils import (
-    DATA_SOURCE_DBAPI_SIGNATURE,
+from snowflake.snowpark._internal.data_source.utils import (
     STATEMENT_PARAMS_DATA_SOURCE,
+    DATA_SOURCE_DBAPI_SIGNATURE,
 )
 from snowflake.snowpark._internal.open_telemetry import open_telemetry_context_manager
 from snowflake.snowpark._internal.telemetry import (
