+++ tests/resources/test_data_source_dir/test_data_source_data.py
@@ -18,6 +18,7 @@ class FakeConnection:
         self.start_index = 0
         self.data = data
         self.schema = schema
+        self.outputtypehandler = None
 
     def cursor(self):
         return self
